Move all C-code to "depsgraph" module
This makes it easier to grab all the code and dump it in any appropriate
directory in the main Blender source code as needed later. It's not clear
yet though which directory exactly these will go to (or whether this will
need to end up as a separate module, or just mixed into blenkernel, since
there will be many cross-references between the two).