Added default handling for relationships on copied nodes
For now, resolution is to clear lists of links when doing base-node copy
(as we have no/little knowledge of what to do with regards to how this
links in with rest of tree).

However, callers of this node should know a bit more
(i.e. when copying node's subtree, at least level n-1 knows
how to solve the links between nodes at level n, though we still have
problems with links which stray to other nodes outside this one...